requirements.txt
Flask==2.2.2
Flask-Login==0.6.2
flask-restx==1.0.3
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
marshmallow==3.19.0
PyJWT==2.6.0
pytest==7.2.0
python-dotenv==0.21.0
SQLAlchemy==1.4.45
WTForms==3.0.1